ci: Move attestation to a new location in workflows
