+++ .github/workflows/package-build.yml
@@ -18,9 +18,6 @@ jobs:
     name: Build package
     runs-on: ubuntu-latest
     environment: package-build
-    permissions:
-      id-token: write
-      attestations: write
     env:
       PACKAGE_NAME: ${{ inputs.package-name || 'tm_devices' }}
     steps:
@@ -29,8 +26,6 @@ jobs:
           fetch-depth: 0
       - uses: hynek/build-and-inspect-python-package@v2.6.0
         id: build-pkg
-        with:
-          attest-build-provenance-github: ${{ !(github.event.pull_request.head.repo.fork || github.event.workflow_call.pull_request.head.repo.fork) && github.actor != 'dependabot[bot]' }}
       - name: Set up Python
         uses: actions/setup-python@v5
         with:
