+++ docs/contributing/add_new_driver.md
@@ -112,7 +112,7 @@ series:
 ```python
 """NewPSU device driver."""
 
-from tm_devices.drivers.pi.power_supplies.psu2200.psu2200 import PSU2200
+from tm_devices.drivers.pi.power_supplies.psu22xx.psu2200 import PSU2200
 
 
 class NewPSU(PSU2200):
