+++ src/tm_devices/drivers/__init__.py
@@ -63,13 +63,13 @@ from tm_devices.drivers.pi.data_acquisition_systems.daq6510 import DAQ6510
 from tm_devices.drivers.pi.digital_multimeters.dmm6500 import DMM6500
 from tm_devices.drivers.pi.digital_multimeters.dmm75xx.dmm7510 import DMM7510
 from tm_devices.drivers.pi.digital_multimeters.dmm75xx.dmm7512 import DMM7512
-from tm_devices.drivers.pi.power_supplies.psu2200.psu2200 import PSU2200
-from tm_devices.drivers.pi.power_supplies.psu2200.psu2220 import PSU2220
-from tm_devices.drivers.pi.power_supplies.psu2200.psu2230 import PSU2230
-from tm_devices.drivers.pi.power_supplies.psu2200.psu2231 import PSU2231
-from tm_devices.drivers.pi.power_supplies.psu2200.psu2231a import PSU2231A
-from tm_devices.drivers.pi.power_supplies.psu2200.psu2280 import PSU2280
-from tm_devices.drivers.pi.power_supplies.psu2200.psu2281 import PSU2281
+from tm_devices.drivers.pi.power_supplies.psu22xx.psu2200 import PSU2200
+from tm_devices.drivers.pi.power_supplies.psu22xx.psu2220 import PSU2220
+from tm_devices.drivers.pi.power_supplies.psu22xx.psu2230 import PSU2230
+from tm_devices.drivers.pi.power_supplies.psu22xx.psu2231 import PSU2231
+from tm_devices.drivers.pi.power_supplies.psu22xx.psu2231a import PSU2231A
+from tm_devices.drivers.pi.power_supplies.psu22xx.psu2280 import PSU2280
+from tm_devices.drivers.pi.power_supplies.psu22xx.psu2281 import PSU2281
 from tm_devices.drivers.pi.source_measure_units.smu24xx.smu2400 import SMU2400
 from tm_devices.drivers.pi.source_measure_units.smu24xx.smu2401 import SMU2401
 from tm_devices.drivers.pi.source_measure_units.smu24xx.smu2410 import SMU2410
