+++ src/tm_devices/drivers/device_driver_mapping.py
@@ -9,13 +9,13 @@ from tm_devices.drivers.pi.data_acquisition_systems.daq6510 import DAQ6510
 from tm_devices.drivers.pi.digital_multimeters.dmm75xx.dmm7510 import DMM7510
 from tm_devices.drivers.pi.digital_multimeters.dmm75xx.dmm7512 import DMM7512
 from tm_devices.drivers.pi.digital_multimeters.dmm6500 import DMM6500
-from tm_devices.drivers.pi.power_supplies.psu2200.psu2200 import PSU2200
-from tm_devices.drivers.pi.power_supplies.psu2200.psu2220 import PSU2220
-from tm_devices.drivers.pi.power_supplies.psu2200.psu2230 import PSU2230
-from tm_devices.drivers.pi.power_supplies.psu2200.psu2231 import PSU2231
-from tm_devices.drivers.pi.power_supplies.psu2200.psu2231a import PSU2231A
-from tm_devices.drivers.pi.power_supplies.psu2200.psu2280 import PSU2280
-from tm_devices.drivers.pi.power_supplies.psu2200.psu2281 import PSU2281
+from tm_devices.drivers.pi.power_supplies.psu22xx.psu2200 import PSU2200
+from tm_devices.drivers.pi.power_supplies.psu22xx.psu2220 import PSU2220
+from tm_devices.drivers.pi.power_supplies.psu22xx.psu2230 import PSU2230
+from tm_devices.drivers.pi.power_supplies.psu22xx.psu2231 import PSU2231
+from tm_devices.drivers.pi.power_supplies.psu22xx.psu2231a import PSU2231A
+from tm_devices.drivers.pi.power_supplies.psu22xx.psu2280 import PSU2280
+from tm_devices.drivers.pi.power_supplies.psu22xx.psu2281 import PSU2281
 from tm_devices.drivers.pi.scopes.tekscope.lpd6 import LPD6
 from tm_devices.drivers.pi.scopes.tekscope.mso2 import MSO2
 from tm_devices.drivers.pi.scopes.tekscope.mso4 import MSO4
